% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biostrext.R
\name{calc_GC}
\alias{calc_GC}
\title{Wrapper for calc_GC_single}
\usage{
calc_GC(x)
}
\arguments{
\item{x}{DNAString or DNAStringSet object}
}
\value{
%GC for single sequence or vector of %GC if DNAStringSet was provided as input
}
\description{
Wrapper for calc_GC_single
}
\examples{
calc_GC(Biostrings::DNAString("ACAGCAGTCA"))
calc_GC(Biostrings::DNAStringSet(c("ACAGCAGTCA","ACGTAGCATCGA","CGATCGACA")))
}
