% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biostrext.R
\name{calc_GC_single}
\alias{calc_GC_single}
\title{Calculates %GC content for a single DNAString object}
\usage{
calc_GC_single(x)
}
\arguments{
\item{x}{DNAString object}
}
\value{
%GC content of provided sequence
}
\description{
Calculates %GC content for a single DNAString object
}
\examples{
calc_GC_single(Biostrings::DNAString("CGAGCACACATCACAC"))
}
