% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biostrext.R
\name{calc_windowed_GC}
\alias{calc_windowed_GC}
\title{Split provided DNAString in chunks and calculate %GC for each chunk}
\usage{
calc_windowed_GC(x, window, overlap)
}
\arguments{
\item{x}{DNAString object}

\item{window}{size of a sequence chunk}

\item{overlap}{size of the overlap between chunks (have to be < window)}
}
\value{
vector of calculated %GC content for each chunk generated
}
\description{
Split provided DNAString in chunks and calculate %GC for each chunk
}
\examples{
calc_windowed_GC(Biostrings::DNAString("CAGTCAGTCGATC"),10,2)
}
