% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biostrext.R
\name{get_max_kmer}
\alias{get_max_kmer}
\title{Calculates maximum kmer for DNA sequence object (DNAString or DNAStringSet)}
\usage{
get_max_kmer(x, k)
}
\arguments{
\item{x}{DNAString or DNAStringSet object}

\item{mer}{kmer size}
}
\value{
named vector of max kmer counts with kmer sequences as names
}
\description{
Calculates maximum kmer for DNA sequence object (DNAString or DNAStringSet)
}
\examples{
get_max_kmer(Biostrings::DNAString("CAGCGATCGTACACGATC"),4)
get_max_kmer(Biostrings::DNAStringSet(c("CAGCGATCGTACACGATC","CAGCGTAGCTAG","CAGCATCACACACAC")),3)
}
