% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biostrext.R
\name{get_max_kmer_single}
\alias{get_max_kmer_single}
\title{Title Get kmer which is most frequent in provided sequence}
\usage{
get_max_kmer_single(x, mer)
}
\arguments{
\item{x}{DNAString object}

\item{mer}{kmer length}
}
\value{
named numeric, with count of given kmer and its sequence as name
}
\description{
Title Get kmer which is most frequent in provided sequence
}
\examples{
get_max_kmer_single(Biostrings::DNAString("CAGCTGCACACATCGTACA"),3)
}
