% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biostrext.R
\name{is_DNAstring}
\alias{is_DNAstring}
\title{Checks if provided object is of class DNAString}
\usage{
is_DNAstring(x)
}
\arguments{
\item{x}{- R object}
}
\value{
True if object is DNAString, false otherwise
}
\description{
Checks if provided object is of class DNAString
}
\examples{
is_DNAstring(Biostrings::DNAString("AGCTAGCA"))
is_DNAstring("AGCTAGCA")
}
