% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biostrext.R
\name{is_DNAstringSet}
\alias{is_DNAstringSet}
\title{Checks if provided object is of class DNAStringSet}
\usage{
is_DNAstringSet(x)
}
\arguments{
\item{x}{- R object}
}
\value{
True if object is DNAString, false otherwise
}
\description{
Checks if provided object is of class DNAStringSet
}
\examples{
is_DNAstringSet(Biostrings::DNAStringSet(c("AGCTAGCA","CACGATC","CGACGAC")))
is_DNAstringSet("AGCTAGCA")
}
