% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biostrext.R
\name{plot_GC_distribution}
\alias{plot_GC_distribution}
\title{Plot %GC distribution across sequence in fragments of given size}
\usage{
plot_GC_distribution(x, window, overlap)
}
\arguments{
\item{x}{DNAString object}

\item{window}{size of the sequence chunk}

\item{overlap}{size of the overlap between chunks}
}
\value{
ggplot2 geom_smooth plot with %GC distribution across sequence
}
\description{
Plot %GC distribution across sequence in fragments of given size
}
\examples{
plot_GC_distribution(Biostrings::DNAString("CAGCTAGCTAGCTAGGCAGCTAGCTAGCTAGCATGCTAGCTAGCTAGCTACGTACGTAGCTACACTAGCTAGCTAGCTAGCTAATAATTATGGCGCGC"),6,2)
}
