% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biostrext.R
\name{plot_kmer_distribution}
\alias{plot_kmer_distribution}
\title{Plots a distribution of given sequence fragment across the whole sequence}
\usage{
plot_kmer_distribution(x, kmer)
}
\arguments{
\item{x}{DNAString object}

\item{kmer}{- string}
}
\value{
plot with distribution of given kmer across sequence
}
\description{
Plots a distribution of given sequence fragment across the whole sequence
}
\examples{
plot_kmer_distribution(Biostrings::DNAString("CAGCTAGCTAGCTAGGCAGCTAGCTAGCTAGCATGCTAGCTAGCTAGCTACGTACGTAGCTACACTAGCTAGCTAGCTAGCTAATAATTATGGCGCGC"),"TAC")
}
