% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biostrext.R
\name{split_in_windows}
\alias{split_in_windows}
\title{Split a DNAString object in smaller chunks}
\usage{
split_in_windows(x, window, overlap)
}
\arguments{
\item{x}{DNAString object}

\item{window}{size of a sequence chunk}

\item{overlap}{size of the overlap between chunks (have to be < window)}
}
\value{
DNAStringSet object
}
\description{
Split a DNAString object in smaller chunks
}
\examples{
split_in_windows(Biostrings::DNAString("ACGTAGCATGCTAGCTACGTAGCTA"),4,1)
}
