% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biostrext.R
\name{is_DNAstringSet}
\alias{is_DNAstringSet}
\title{#' Produce error if is_DNAstring is false
#'
#' @param call is_DNAstring invocation
#' @param env
#'
#' @return Error if is_DNAstring call is false
#'
on_failure(is_DNAstring) <- function(call, env) {
  paste0(deparse(call$x), " is not a DNAString object")
}
Checks if provided object is of class DNAStringSet}
\usage{
is_DNAstringSet(x)
}
\arguments{
\item{x}{- R object}
}
\value{
True if object is DNAString, false otherwise
}
\description{
#' Produce error if is_DNAstring is false
#'
#' @param call is_DNAstring invocation
#' @param env
#'
#' @return Error if is_DNAstring call is false
#'
on_failure(is_DNAstring) <- function(call, env) {
  paste0(deparse(call$x), " is not a DNAString object")
}
Checks if provided object is of class DNAStringSet
}
\examples{
is_DNAstringSet(Biostrings::DNAStringSet(c("AGCTAGCA","CACGATC","CGACGAC")))
is_DNAstringSet("AGCTAGCA")
}
